banner/requirements.txt
Flask==3.0.3
Pillow==10.3.0
mcstatus==11.1.1
